GradientDescent

Description

|

Documentation

sigmoid :: Double -> Double

Sigmoid Function

sigmoidPrime :: Double -> Double

Sigmoid Prime Function

matrixSigmoid :: Matrix Double -> Matrix Double

Apply Sigmoid Function for every element of matrix

matrixSigmoidPrime :: Matrix Double -> Matrix Double

Apply Sigmoid Prime Function for every element of matrix

getOutcomes

Arguments

:: Net
Neural Network

-> [(Matrix Double, Int)]
List of Test Images with their corresponding labels

-> [([Float], (Int, Int))]
Returns List of Test Images with their corresponding label and predictions

Get Neural Network predictions

gradientDescent

Arguments

:: Net
Neural Network to train

-> [(Matrix Double, Matrix Double)]
List of Training Images with column vectorized labels

-> Int
Number of training iterations

-> Int
Size of Image batches

-> Double
Alpha

-> [(Matrix Double, Int)]
List of Test Images with Labels

-> IO Net

Train Neural Network